third_party/blink/web_tests/platform/win/external/wpt/editing/run/delete_1001-2000-expected.txt

This is a testharness.js-based test.
Found 32 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["delete",""]] "<p>foo<table><tr><td>bar</table><p>[]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><table><tbody><tr><td>bar</td></tr></tbody></table><p>baz</p>" but got "<p>foo</p><table><tbody><tr><td>bar</td></tr></tbody></table>baz"
[FAIL] [["delete",""]] "foo<br><table><tr><td>[]bar</table>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<table><tbody><tr><td>bar</td></tr></tbody></table>baz" but got "foo<br><table><tbody><tr><td>bar</td></tr></tbody></table>baz"
[FAIL] [["delete",""]] "<p>foo<table><tr><td>bar<br></table><p>[]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><table><tbody><tr><td>bar<br></td></tr></tbody></table><p>baz</p>" but got "<p>foo</p><table><tbody><tr><td>bar<br></td></tr></tbody></table>baz"
[FAIL] [["delete",""]] "<table><tr><td>foo<br><td>[]bar</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td>foo</td><td>bar</td></tr></tbody></table>" but got "<table><tbody><tr><td>foo<br></td><td>bar</td></tr></tbody></table>"
[FAIL] [["delete",""]] "<table><tr><td>foo<br><tr><td>[]bar</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td>foo</td></tr><tr><td>bar</td></tr></tbody></table>" but got "<table><tbody><tr><td>foo<br></td></tr><tr><td>bar</td></tr></tbody></table>"
[FAIL] [["delete",""]] "<p>foo<table><tr><td>bar<br><br></table><p>[]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><table><tbody><tr><td>bar<br><br></td></tr></tbody></table><p>baz</p>" but got "<p>foo</p><table><tbody><tr><td>bar<br><br></td></tr></tbody></table>baz"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "foo<ol><li>[]bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div>bar</div><ol><li>baz</li></ol>" but got "foobar<ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "foo<ol><li>[]bar<li>baz</ol>" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "foo<ol><li>[]bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<p>bar</p><ol><li>baz</li></ol>" but got "foobar<ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "foo<br><ol><li>[]bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div>bar</div><ol><li>baz</li></ol>" but got "foobar<ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "foo<br><ol><li>[]bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<p>bar</p><ol><li>baz</li></ol>" but got "foobar<ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "foo<br><br><ol><li>[]bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br><br><div>bar</div><ol><li>baz</li></ol>" but got "foo<br><ol><li>bar</li><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "foo<br><br><ol><li>[]bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br><br><p>bar</p><ol><li>baz</li></ol>" but got "foo<br><ol><li>bar</li><li>baz</li></ol>"
[FAIL] [["delete",""]] "<ol><li>foo<li>[]bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<br>bar</li></ol>" but got "<ol><li>foobar</li></ol>"
[FAIL] [["delete",""]] "<ol><li>foo<br><li>[]bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<br>bar</li></ol>" but got "<ol><li>foobar</li></ol>"
[FAIL] [["delete",""]] "<ol><li>foo<br><br><li>[]bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<br><br>bar</li></ol>" but got "<ol><li>foo<br>bar</li></ol>"
[FAIL] [["delete",""]] "<ol><li>foo<li>[]bar<br>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<br>bar<br>baz</li></ol>" but got "<ol><li>foobar</li><li>baz</li></ol>"
[FAIL] [["delete",""]] "<ol><li>foo<br>bar<li>[]baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<br>bar<br>baz</li></ol>" but got "<ol><li>foo<br>barbaz</li></ol>"
[FAIL] [["delete",""]] "<ol><li><p>foo<li>[]bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><p>foo</p>bar</li></ol>" but got "<ol><li><p>foobar</p></li></ol>"
[FAIL] [["delete",""]] "<ol><li>foo<li><p>[]bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<p>bar</p></li></ol>" but got "<ol><li>foobar</li></ol>"
[FAIL] [["delete",""]] "<ol><li><p>foo<li><p>[]bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><p>foo</p><p>bar</p></li></ol>" but got "<ol><li><p>foobar</p></li></ol>"
[FAIL] [["delete",""]] "<ol><li>foo<ul><li>[]bar</ul></ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>bar</li></ol>" but got "<ol><li>foobar</li></ol>"
[FAIL] [["delete",""]] "foo<ol><ol><li>[]bar</ol></ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<ol><li>bar</li></ol>" but got "foobar"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "foo<div><ol><li>[]bar</ol></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div><div>bar</div></div>" but got "foobar"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "foo<div><ol><li>[]bar</ol></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div><p>bar</p></div>" but got "foobar"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "foo<dl><dt>[]bar<dd>baz</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div>bar</div><dl><dd>baz</dd></dl>" but got "foobar<dl><dd>baz</dd></dl>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "foo<dl><dt>[]bar<dd>baz</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<p>bar</p><dl><dd>baz</dd></dl>" but got "foobar<dl><dd>baz</dd></dl>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "foo<dl><dd>[]bar</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div>bar</div>" but got "foobar"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "foo<dl><dd>[]bar</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<p>bar</p>" but got "foobar"
[FAIL] [["delete",""]] "<dl><dt>foo<dd>[]bar</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo<br>bar</dt></dl>" but got "<dl><dt>foobar</dt></dl>"
[FAIL] [["delete",""]] "<dl><dt>foo<dt>[]bar<dd>baz</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo<br>bar</dt><dd>baz</dd></dl>" but got "<dl><dt>foobar</dt><dd>baz</dd></dl>"
[FAIL] [["delete",""]] "<dl><dt>foo<dd>bar<dd>[]baz</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dd>bar<br>baz</dd></dl>" but got "<dl><dt>foo</dt><dd>barbaz</dd></dl>"
Harness: the test ran to completion.
